Fully functioning assessment methods
Still need to hookup the elevator proxy with the manager...and figure
out how to actually assess the elevator manager

+++ src/people/ISchedule.java
@@ -2,5 +2,5 @@ package people;
 
 public interface ISchedule {
 	public abstract void initSchedule(IPerson[] p) throws Exception;
-	public abstract IElevatorTrip[] getRequestsAtTime(double currentTime);
+	public abstract IElevatorTrip[] getRequestsAtTime(int currentTime);
 }
